requirements.txt
chime==0.7.0
deep_sort_realtime==1.3.2
deep_translator==1.11.4
elevenlabs==1.2.2
joblib==1.4.2
keras==3.3.3
llama_index==0.10.43
numpy==1.26.4
openai==1.33.0
opencv_python==4.9.0.80
Pillow==10.3.0
pvporcupine==3.0.2
PyAudio==0.2.14
scipy==1.13.1
SpeechRecognition==3.10.4
torch==2.2.2
transformers==4.41.0
ultralytics==8.2.15
